+++ src/youwol/backends/assets/routers/__init__.py
@@ -1,2 +1,3 @@
 # relative
-from .files import *
+from .files import router as router_files
+from .images import router as router_images
